Merger & placer modification implemented succesfully. Integrating into the layout phase.

+++ gila-multi/src/unipg/gila/multi/placers/PlacerCoordinateDelivery.java
@@ -11,6 +11,7 @@ import java.util.Iterator;
 import org.apache.giraph.edge.Edge;
 import org.apache.giraph.graph.Vertex;
 import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.IntWritable;
 
 import unipg.gila.common.datastructures.messagetypes.LayoutMessage;
 import unipg.gila.multi.MultiScaleComputation;
@@ -28,16 +29,16 @@ public class PlacerCoordinateDelivery extends MultiScaleComputation<AstralBodyCo
 	 */
 	@Override
 	protected void vertexInLayerComputation(
-			Vertex<LayeredPartitionedLongWritable, AstralBodyCoordinateWritable, FloatWritable> vertex,
+			Vertex<LayeredPartitionedLongWritable, AstralBodyCoordinateWritable, IntWritable> vertex,
 			Iterable<LayoutMessage> msgs) throws IOException {
 		Iterator<LayoutMessage> ms = msgs.iterator();
 		if(!ms.hasNext() || vertex.getValue().isSun())
 			return;
 		Long currentVertexID = vertex.getId().getId();
 		HashMap<Long, LayeredPartitionedLongWritable> myNeighbors = new HashMap<Long, LayeredPartitionedLongWritable>();
-		Iterator<Edge<LayeredPartitionedLongWritable, FloatWritable>> edges = vertex.getEdges().iterator();
+		Iterator<Edge<LayeredPartitionedLongWritable, IntWritable>> edges = vertex.getEdges().iterator();
 		while(edges.hasNext()){
-			Edge<LayeredPartitionedLongWritable, FloatWritable> currentEdge = edges.next();
+			Edge<LayeredPartitionedLongWritable, IntWritable> currentEdge = edges.next();
 			myNeighbors.put(currentEdge.getTargetVertexId().getId(), currentEdge.getTargetVertexId());
 		}
 		while(ms.hasNext()){
